Make a mess
Here is the first iteration of the literate config!

Let's see how it holds up. It kind of seems like everything is split
up, but it's good to see the monolithic config be torn down in favor
of something more modular. There are still some quirks to work out,
like how to choose what packages and themes to actually load and use
out of what is in the subdirectories, but the data-driven approach
of "what you see is what you get" is never a bad thing.