[GARRY] In the beginning...
This is the full machine-specific config for Garry, my main desktop
computer. It includes a mode line that I wrote myself, custom startup
screen (the ASCII lambda), and general bits and bobs that have
accumulated over my time using Emacs on this machine.

The idea is to store each machine I use Emacs on with it's own
specific branch, while keeping the main branch as a way to jump-start
a new machine into something akin to an Emacs I am used to.